Move onAuthFlowComplete to new AuthenticationUtilities file.
